+++ txnbuild/transaction.go
@@ -189,6 +189,17 @@ func marshallBase64(e xdr.TransactionEnvelope, signatures []xdr.DecoratedSignatu
 	return base64.StdEncoding.EncodeToString(binary), nil
 }
 
+func marshallBase64Bytes(e xdr.TransactionEnvelope, signatures []xdr.DecoratedSignature) ([]byte, error) {
+	binary, err := marshallBinary(e, signatures)
+	if err != nil {
+		return nil, errors.Wrap(err, "failed to get XDR bytestring")
+	}
+
+	encoded := make([]byte, base64.StdEncoding.EncodedLen(len(binary)))
+	base64.StdEncoding.Encode(encoded, binary)
+	return encoded, nil
+}
+
 // Transaction represents a Stellar transaction. See
 // https://www.stellar.org/developers/guides/concepts/transactions.html
 // A Transaction may be wrapped by a FeeBumpTransaction in which case
@@ -346,9 +357,34 @@ func (t *Transaction) MarshalBinary() ([]byte, error) {
 	return marshallBinary(t.envelope, t.Signatures())
 }
 
+// MarshalText returns the base64 XDR representation of the transaction envelope.
+func (t *Transaction) MarshalText() ([]byte, error) {
+	return marshallBase64Bytes(t.envelope, t.Signatures())
+}
+
+// UnmarshalText consumes into the value the base64 XDR representation of the
+// transaction envelope.
+func (t *Transaction) UnmarshalText(b []byte) error {
+	gtx, err := TransactionFromXDR(string(b))
+	if err != nil {
+		return err
+	}
+	tx, ok := gtx.Transaction()
+	if !ok {
+		return errors.New("transaction envelope unmarshaled into FeeBumpTransaction is not a fee bump transaction")
+	}
+	*t = *tx
+	return nil
+}
+
 // Base64 returns the base 64 XDR representation of the transaction envelope.
 func (t *Transaction) Base64() (string, error) {
 	return marshallBase64(t.envelope, t.Signatures())
+}
+
+// ToGenericTransaction creates a GenericTransaction containing the Transaction.
+func (t *Transaction) ToGenericTransaction() *GenericTransaction {
+	return &GenericTransaction{simple: t}
 }
 
 // ClaimableBalanceID returns the claimable balance ID for the operation at the given index within the transaction.
@@ -506,9 +542,36 @@ func (t *FeeBumpTransaction) MarshalBinary() ([]byte, error) {
 	return marshallBinary(t.envelope, t.Signatures())
 }
 
+// MarshalText returns the base64 XDR representation of the transaction
+// envelope.
+func (t *FeeBumpTransaction) MarshalText() ([]byte, error) {
+	return marshallBase64Bytes(t.envelope, t.Signatures())
+}
+
+// UnmarshalText consumes into the value the base64 XDR representation of the
+// transaction envelope.
+func (t *FeeBumpTransaction) UnmarshalText(b []byte) error {
+	gtx, err := TransactionFromXDR(string(b))
+	if err != nil {
+		return err
+	}
+	fbtx, ok := gtx.FeeBump()
+	if !ok {
+		return errors.New("transaction envelope unmarshaled into Transaction is not a transaction")
+	}
+	*t = *fbtx
+	return nil
+}
+
 // Base64 returns the base 64 XDR representation of the transaction envelope.
 func (t *FeeBumpTransaction) Base64() (string, error) {
 	return marshallBase64(t.envelope, t.Signatures())
+}
+
+// ToGenericTransaction creates a GenericTransaction containing the
+// FeeBumpTransaction.
+func (t *FeeBumpTransaction) ToGenericTransaction() *GenericTransaction {
+	return &GenericTransaction{feeBump: t}
 }
 
 // InnerTransaction returns the Transaction which is wrapped by
@@ -538,6 +601,29 @@ func (t GenericTransaction) Transaction() (*Transaction, bool) {
 // can be unpacked into a FeeBumpTransaction.
 func (t GenericTransaction) FeeBump() (*FeeBumpTransaction, bool) {
 	return t.feeBump, t.feeBump != nil
+}
+
+// MarshalText returns the base64 XDR representation of the transaction
+// envelope.
+func (t *GenericTransaction) MarshalText() ([]byte, error) {
+	if tx, ok := t.Transaction(); ok {
+		return tx.MarshalText()
+	}
+	if fbtx, ok := t.FeeBump(); ok {
+		return fbtx.MarshalText()
+	}
+	return nil, errors.New("unable to marshal empty GenericTransaction")
+}
+
+// UnmarshalText consumes into the value the base64 XDR representation of the
+// transaction envelope.
+func (t *GenericTransaction) UnmarshalText(b []byte) error {
+	gtx, err := TransactionFromXDR(string(b))
+	if err != nil {
+		return err
+	}
+	*t = *gtx
+	return nil
 }
 
 type TransactionFromXDROption int