+++ txnbuild/CHANGELOG.md
@@ -6,6 +6,9 @@ file.  This project adheres to [Semantic Versioning](http://semver.org/).
 
 ## Unreleased
 
+* GenericTransaction, Transaction, and FeeBumpTransaction now implement
+encoding.TextMarshaler and encoding.TextUnmarshaler.
+
 ## [v7.1.1](https://github.com/stellar/go/releases/tag/horizonclient-v7.1.1) - 2021-06-25
 
 ### Bug Fixes